Separate folders for MVC Layout, more game code.
The code that i added in Check fit itterates over the cells in the map and sees if any of them are X, if they are then it returns false.

Variables for game logic

after i could draw the map, Im trying to draw the pieces. Draw Current Piece does just that.  iterates over all cells in tetromino and gets the piece then it sees if it is rotated and at what degree using the equations up top, then draws that to the screen.

included files in diffrent folders for future modularization